Timeline: fix line with dynamic items
When inserting new timeline-items on a timeline it could occur that the
item that was previously marked as the last is actually not the last
item anymore. This would result in a broken timeline.

The problem, as I see it, is that the parent component's has no trigger
to run updateChildren again (which in theory could resole this).
However, I decided to solve the issue in a simpler way. CSS already
knows for sure which component is the last; it can simply hide the last
line for us and thus we can remove a bit of the interdependency between
the two components.